CLOUD-520 revise Gcp and Azure cloud resource related logic
